refactor: rename padded dense to jagged module
